ボタンクライアント関数を分離
