+++ src/robot-server.ts
@@ -21,6 +21,7 @@ import axios, { Method } from "axios"
 import { selectEngine } from "./speech"
 import { Talk } from "./talk"
 import { ButtonClient } from "./button-client"
+import { BottonModule } from "./button-module"
 import { RobotDB } from "./robot-db"
 import { router as googleRouter } from "./google-router"
 import * as session from "express-session"
@@ -144,170 +145,7 @@ const Dora = require("dora")
 const dora = new Dora()
 const utils = require("./utils")
 
-dora.loadModule("button", function (DORA, config) {
-  function connect(node, options) {
-    const params = options.split("/")
-    if (params.length < 1 || params === "") {
-      throw new Error("ホスト名がありません。")
-    }
-    node.on("input", async function (msg) {
-      let host = params[0]
-      let port = null
-      if (host.indexOf(":") > 0) {
-        const h = host.split(":")
-        host = h[0]
-        port = h[1]
-      }
-      const name = params.length > 1 ? params[1] : null
-      const team = params.length > 2 ? params[2] : null
-      buttonClient.emit("open-slave", { host, name, team, port })
-      node.send(msg)
-    })
-  }
-  DORA.registerType("connect", connect)
-
-  function close(node, options) {
-    const params = options.split("/")
-    if (params.length < 1 || params === "") {
-      throw new Error("ホスト名がありません。")
-    }
-    node.on("input", async function (msg) {
-      const host = params[0]
-      buttonClient.emit("close-slave", { host })
-      node.send(msg)
-    })
-  }
-  DORA.registerType("close", close)
-
-  function allBlink(node, options) {
-    node.on("input", async function (msg) {
-      buttonClient.emit("all-blink", {})
-      node.send(msg)
-    })
-  }
-  DORA.registerType("led-all-blink", allBlink)
-
-  function allOn(node, options) {
-    node.on("input", async function (msg) {
-      buttonClient.emit("all-on", {
-        bright: 1,
-      })
-      node.send(msg)
-    })
-  }
-  DORA.registerType("led-all-on", allOn)
-
-  function allOff(node, options) {
-    node.on("input", async function (msg) {
-      buttonClient.emit("all-off", {})
-      node.send(msg)
-    })
-  }
-  DORA.registerType("led-all-off", allOff)
-
-  function ledOn(node, options) {
-    node.on("input", async function (msg) {
-      buttonClient.emit("one", {
-        name: msg.button.name,
-        bright: 1,
-      })
-      node.send(msg)
-    })
-  }
-  DORA.registerType("led-on", ledOn)
-
-  function sound(node, options) {
-    let isTemplated = (options || "").indexOf("{{") != -1
-    node.on("input", async function (msg) {
-      const socket = buttonClient.socket(msg.button.name)
-      if (socket) {
-        let message = options
-        if (isTemplated) {
-          message = DORA.utils.mustache.render(message, msg)
-        }
-        socket.emit("sound-command", { sound: message })
-      }
-      node.send(msg)
-    })
-  }
-  DORA.registerType("sound", sound)
-
-  function soundAll(node, options) {
-    let isTemplated = (options || "").indexOf("{{") != -1
-    node.on("input", async function (msg) {
-      let message = options
-      if (isTemplated) {
-        message = DORA.utils.mustache.render(message, msg)
-      }
-      buttonClient.emit("sound", { sound: message })
-      node.send(msg)
-    })
-  }
-  DORA.registerType("sound-all", soundAll)
-
-  function speechToText(node, options) {
-    node.nextLabel(options)
-    node.on("input", async function (msg) {
-      const socket = buttonClient.socket(msg.button.name)
-      if (socket) {
-        const params = {
-          timeout: 30000,
-          sensitivity: "keep",
-        }
-        if (typeof msg.timeout !== "undefined") {
-          params.timeout = msg.timeout
-        }
-        if (typeof msg.sensitivity !== "undefined") {
-          params.sensitivity = msg.sensitivity
-        }
-        node.recording = true
-        socket.emit("speech-to-text", params, (res) => {
-          if (!node.recording) return
-          node.recording = false
-          if (res == "[timeout]") {
-            msg.payload = "timeout"
-            node.send([msg, null])
-          } else if (res == "[canceled]") {
-            msg.payload = "canceled"
-            node.send([msg, null])
-          } else {
-            if (res.button) {
-              msg.payload = "button"
-              msg.button = res
-              delete res.button
-              node.send([msg, null])
-            } else if (res.speechRequest) {
-              msg.speechRequest = true
-              msg.payload = res.payload
-              msg.speechText = msg.payload
-              msg.topicPriority = 0
-              node.send([null, msg])
-            } else if (typeof res === "object") {
-              ;(msg.languageCode = res.languageCode),
-                (msg.alternativeLanguageCodes = res.alternativeLanguageCodes),
-                (msg.confidence = res.confidence)
-              msg.payload = res.transcript
-              msg.speechText = msg.payload
-              msg.topicPriority = 0
-              delete msg.speechRequest
-              node.next(msg)
-            } else {
-              msg.payload = res
-              msg.speechText = msg.payload
-              msg.topicPriority = 0
-              delete msg.speechRequest
-              node.send([null, msg])
-            }
-          }
-        })
-      } else {
-        msg.payload = "timeout"
-        node.send([msg, null])
-      }
-    })
-  }
-  DORA.registerType("speech-to-text", speechToText)
-})
+dora.loadModule("button", BottonModule(buttonClient))
 
 dora.request = async function (command, options = null, params = null) {
   let len = 0
